work on authorization

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
     title: req.body.title,
     content: req.body.content,
     // multer provides .file
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   //save provided by mongoose
   post.save().then( createdPost => {
@@ -109,21 +110,34 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then( result => {
-    res.status(200).json({
-      message: 'Update successful!'
-    });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then( result => {
+    if (result.nModified > 0) {
+      res.status(200).json({
+        message: 'Update successful!'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized!'
+      });
+    }
+
   });
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then( result => {
-    res.status(200).json({
-      message: 'Post deleted!'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then( result => {
+    if (result.n > 0) {
+      res.status(200).json({
+        message: 'Deletion successful!'
+      });
+    } else {
+      res.status(401).json({
+        message: 'Not authorized!'
+      });
+    }
   });
 });
